requirements.txt
python-dotenv==1.0.0
langchain-community==0.2.1
langchain-openai==0.1.0
langchain-qdrant==0.1.0
qdrant-client==1.4.3
